requirements.txt
enum34
future
numpy>=1.8.2
scipy>=0.17.0
matplotlib
pyparma>=0.3.3
pycddlib>=1.0.0
cvxopt>=1.1.4
